refactor(commands): move abstract commands
move abstract command files to `lib/Structures/commands`
